Switch from DL4J to TensorFlow

+++ src/main/kotlin/com/jetbrains/deepbugs/utils/Mapping.kt
@@ -3,5 +3,5 @@ package com.jetbrains.deepbugs.utils
 import com.beust.klaxon.JsonObject
 
 class Mapping(private val json: JsonObject) {
-    fun get(fieldName: String) = json.array<Double>(fieldName)
+    fun get(fieldName: String) = json.array<Float>(fieldName)
 }
